+++ .storybook/preview.js
@@ -6,7 +6,6 @@ import Cookies from "../src/nationalarchives/lib/cookies.mjs";
 
 document.documentElement.classList.add(
   "tna-template",
-  "tna-template--light-theme",
   "tna-template--yellow-accent",
 );
 if (window.self !== window.top) {